refactor(dev): split cmd/dev/main.go into focused files (setup, mint, transfer, validate)
